napravljen LogOutButton komponeta sa Modalom za SIdeBarNavigation i updejtan SidebarNavigation
